requirements.txt
faster-whisper
google-genai
mediapipe
numpy
opencv-python
python-dotenv
youtube-transcript-api
yt-dlp